requirements.txt
langchain-core
langchain-openai
streamlit
langchain-text-splitters
langchain-community
langchain>=0.1.0
openai>=1.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1
python-dotenv>=1.0.0
pypdf>=3.9.0
fastapi
uvicorn
